refactor calibration to use a base class
