+++ scos_actions/calibration/sensor_calibration.py
@@ -4,25 +4,15 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Union
 
-from scos_actions.signal_processing.calibration import CalibrationException
+from scos_actions.calibration.interfaces.calibration import Calibration
 
 logger = logging.getLogger(__name__)
 
 
 @dataclass
-class SensorCalibration:
+class SensorCalibration(Calibration):
     last_calibration_datetime: str
-    calibration_parameters: List[str]
-    calibration_data: dict
     clock_rate_lookup_by_sample_rate: List[Dict[str, float]]
-    is_default: bool
-    file_path: Path
-
-    def __post_init__(self):
-        # Convert key names in calibration_data to strings
-        # This means that formatting will always match between
-        # native types provided in Python and data loaded from JSON
-        self.calibration_data = json.loads(json.dumps(self.calibration_data))
 
     def get_clock_rate(self, sample_rate: Union[float, int]) -> Union[float, int]:
         """Find the clock rate (Hz) using the given sample_rate (samples per second)"""
@@ -30,25 +20,6 @@ class SensorCalibration:
             if mapping["sample_rate"] == sample_rate:
                 return mapping["clock_frequency"]
         return sample_rate
-
-    def get_calibration_dict(self, cal_params: List[Union[float, int, bool]]) -> dict:
-        """
-        Get calibration data closest to the specified parameter values.
-
-        :param cal_params: List of calibration parameter values. For example,
-            if ``calibration_parameters`` are ``["sample_rate", "gain"]``,
-            then the input to this method could be ``["15360000.0", "40"]``.
-        :return: The calibration data corresponding to the input parameter values.
-        """
-
-        cal_data = self.calibration_data
-        for i, setting_value in enumerate(cal_params):
-            setting = self.calibration_parameters[i]
-            logger.debug(f"Looking up calibration for {setting} at {setting_value}")
-            cal_data = filter_by_parameter(cal_data, setting_value)
-        logger.debug(f"Got calibration data: {cal_data}")
-
-        return cal_data
 
     def update(
         self,
@@ -76,32 +47,14 @@ class SensorCalibration:
         :param file_path: File path for saving the updated calibration data.
         :raises Exception:
         """
-        cal_data = self.calibration_data
+        # Get existing calibration data entry which will be updated
+        data_entry = self._retrieve_data_to_update(params)
+
+        # Update last calibration datetime
         self.last_calibration_datetime = calibration_datetime_str
-        if len(self.calibration_parameters) == 0:
-            self.calibration_parameters = list(params.keys())
-        # Ensure all required calibration parameters were used
-        elif not set(params.keys()) >= set(self.calibration_parameters):
-            raise Exception(
-                "Not enough parameters specified to update calibration.\n"
-                + f"Required parameters are {self.calibration_parameters}"
-            )
 
-        # Get calibration entry by parameters used
-        for parameter in self.calibration_parameters:
-            value = str(params[parameter]).lower()
-            logger.debug(f"Updating calibration at {parameter} = {value}")
-            try:
-                cal_data = cal_data[value]
-            except KeyError:
-                logger.debug(
-                    f"Creating required calibration data field for {parameter} = {value}"
-                )
-                cal_data[value] = {}
-                cal_data = cal_data[value]
-
-        # Update calibration data
-        cal_data.update(
+        # Update calibration data entry (updates entry in self.calibration_data)
+        data_entry.update(
             {
                 "datetime": calibration_datetime_str,
                 "gain": gain_dB,
@@ -120,95 +73,22 @@ class SensorCalibration:
         with open(self.file_path, "w") as outfile:
             outfile.write(json.dumps(cal_dict))
 
+    # @classmethod
+    # def from_json(cls, fname: Path, is_default: bool):
+    #     """
+    #     Load a sensor calibration from a JSON file.
 
-def load_from_json(fname: Path, is_default: bool) -> SensorCalibration:
-    """
-    Load a calibration from a JSON file.
+    #     The JSON file must contain top-level fields:
+    #         ``calibration_parameters``
+    #         ``calibration_data``
+    #         ``last_calibration_datetime``
+    #         ``clock_rate_lookup_by_sample_rate``
 
-    The JSON file must contain top-level fields:
-        ``last_calibration_datetime``
-        ``calibration_parameters``
-        ``calibration_data``
-        ``clock_rate_lookup_by_sample_rate``
-
-    :param fname: The ``Path`` to the JSON calibration file.
-    :param is_default: If True, the loaded calibration file
-        is treated as the default calibration file.
-    :raises Exception: If the provided file does not include
-        the required keys.
-    :return: The ``Calibration`` object generated from the file.
-    """
-    with open(fname) as file:
-        calibration = json.load(file)
-    # Check that the required fields are in the dict
-    required_keys = {
-        "last_calibration_datetime",
-        "calibration_data",
-        "clock_rate_lookup_by_sample_rate",
-        "calibration_parameters",
-    }
-
-    if not calibration.keys() >= required_keys:
-        raise Exception(
-            "Loaded calibration dictionary is missing required fields."
-            + f"Existing fields: {set(calibration.keys())}\n"
-            + f"Required fields: {required_keys}\n"
-        )
-    # Create and return the Calibration object
-    return SensorCalibration(
-        calibration["last_calibration_datetime"],
-        calibration["calibration_parameters"],
-        calibration["calibration_data"],
-        calibration["clock_rate_lookup_by_sample_rate"],
-        is_default,
-        fname,
-    )
-
-
-def filter_by_parameter(calibrations: dict, value: Union[float, int, bool]) -> dict:
-    """
-    Select a certain element by the value of a top-level key in a dictionary.
-
-    This method should be recursively called to select calibration
-    data matching a set of calibration parameters. The ordering of
-    nested dictionaries should match the ordering of the required
-    calibration parameters in the calibration file.
-
-    If ``value`` is a float or bool, ``str(value).lower()`` is used
-    as the dictionary key. If ``value`` is an int, and the previous
-    approach does not work, ``str(float(value))`` is attempted. This
-    allows for value ``1`` to match a key ``"1.0"``.
-
-    :param calibrations: Calibration data dictionary.
-    :param value: The parameter value for filtering. This value should
-        exist as a top-level key in ``calibrations``.
-    :raises CalibrationException: If ``value`` cannot be matched to a
-        top-level key in ``calibrations``, or if ``calibrations`` is not
-        a dict.
-    :return: The value of ``calibrations[value]``, which should be a dict.
-    """
-    try:
-        filtered_data = calibrations.get(str(value).lower(), None)
-        if filtered_data is None and isinstance(value, int):
-            # Try equivalent float for ints, i.e., match "1.0" to 1
-            filtered_data = calibrations.get(str(float(value)), None)
-        if filtered_data is None and isinstance(value, float) and value.is_integer():
-            # Check for, e.g., key '25' if value is '25.0'
-            filtered_data = calibrations.get(str(int(value)), None)
-        if filtered_data is None:
-            raise KeyError
-        else:
-            return filtered_data
-    except AttributeError as e:
-        # calibrations does not have ".get()"
-        # Generally means that calibrations is None or not a dict
-        msg = f"Provided calibration data is not a dict: {calibrations}"
-        raise CalibrationException(msg)
-    except KeyError as e:
-        msg = (
-            f"Could not locate calibration data at {value}"
-            + f"\nAttempted lookup using key '{str(value).lower()}'"
-            + f"{f'and {float(value)}' if isinstance(value, int) else ''}"
-            + f"\nUsing calibration data: {calibrations}"
-        )
-        raise CalibrationException(msg)
+    #     :param fname: The ``Path`` to the JSON calibration file.
+    #     :param is_default: If True, the loaded calibration file
+    #         is treated as the default calibration file.
+    #     :raises Exception: If the provided file does not include
+    #         the required keys.
+    #     :return: The ``Calibration`` object generated from the file.
+    #     """
+    #     return super().from_json(fname, is_default)
