+++ scos_actions/signal_processing/calibration.py
@@ -3,7 +3,9 @@ from typing import Optional, Tuple
 
 import numpy as np
 from its_preselector.preselector import Preselector
+from numpy.typing import NDArray
 from scipy.constants import Boltzmann
+from scos_actions.calibration.utils import CalibrationException
 from scos_actions.signal_processing.unit_conversion import (
     convert_celsius_to_fahrenheit,
     convert_celsius_to_kelvins,
@@ -15,16 +17,9 @@ from scos_actions.signal_processing.unit_conversion import (
 logger = logging.getLogger(__name__)
 
 
-class CalibrationException(Exception):
-    """Basic exception handling for calibration functions."""
-
-    def __init__(self, msg):
-        super().__init__(msg)
-
-
 def y_factor(
-    pwr_noise_on_watts: np.ndarray,
-    pwr_noise_off_watts: np.ndarray,
+    pwr_noise_on_watts: NDArray,
+    pwr_noise_off_watts: NDArray,
     enr_linear: float,
     enbw_hz: float,
     temp_kelvins: float = 300.0,
